requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
deepl==1.11.0
et-xmlfile==1.1.0
google-api-core==2.10.2
google-auth==2.13.0
google-cloud-speech==2.16.2
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
idna==3.4
numpy==1.23.4
openai==0.24.0
openpyxl==3.0.10
pandas==1.5.1
pandas-stubs==1.5.0.221012
proto-plus==1.22.1
protobuf==4.21.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
rsa==4.9
six==1.16.0
tqdm==4.64.1
types-pytz==2022.5.0.0
typing_extensions==4.4.0
urllib3==1.26.12
